backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
sqlalchemy==2.0.35
pydantic-settings==2.6.0
python-multipart==0.0.12
httpx==0.27.2
